Added JNI glue for getifaddrs native call
Also included a few utility methods for checking private addresses, some of which
are kinda copypasted but since we can't really make them generic to be great for getting
rid of both repetitive code AND make them great for the caller (by e.g. inlining the array in the call itself)
let's just leave them like this since this is a free demo and not production code after all.