BUGFIX: if no public addresses are found, loopback device shouldn't be the only returned IPv4 choice
no longer consider loopback device as an actual valid IP address over other non-public addresses.
while it's not specifically stated in the task description, product-wise it makes sense so that's
why i'm doing it.

Also because it fits the fallback part (point C) of the problem:
a. If an IPv6 address is available, that is from the global unicast range - as per this document, return that address.
b. If an IPv4 address is available that is public - i.e. not a “private address” according to the table here, or “link local” according to the specification here, return that address.
c. Otherwise, return any one of the IPv4 addresses available.