Final polish (#3)
* Add INTERNET permission to manifest

* fix native-lib; ipv4 private addresses should be IGNORED instead of RETURNED

Title says it all, reverse the condition to !isIPv4Private

* QoL improvements; button spam prevented; sending of request seems to work fine

title says it all, button click listener has been guarded and extracted and button can't be spammed

* A bit more love for the SendRequest

As error handling got a bit more complicated, the SendRequest.ResponseResult constructor
was made private since i can't have multiple places instantiating it potentially wrongly...

two new static methods - successfulResponse() and unsuccessfulResponse() are provided which
instantiate it properly; one with setting the expected result, and the other setting the exception

EDIT: also, added actual response body to the ResponseResult.

* Fix native lib to break out of loop once first acceptable address is found to avoid overwriting it

Still seems to only return 127.0.0.1 however so... i'm skeptical.

* Added logging of response body and fault reason

* Fixes across the board

- implemented the overlay fragment
- tied in the overlay fragment with the actual result
- added result showing to the main activity as well

During testing, some bugs were encountered, but eventually sorted out
as well, so the test scenario of shooting five consecutive requests
and having overlays showing up for each of them - works as expected.

* BUGFIX: if no public addresses are found, loopback device shouldn't be the only returned IPv4 choice

no longer consider loopback device as an actual valid IP address over other non-public addresses.
while it's not specifically stated in the task description, product-wise it makes sense so that's
why i'm doing it.

Also because it fits the fallback part (point C) of the problem:
a. If an IPv6 address is available, that is from the global unicast range - as per this document, return that address.
b. If an IPv4 address is available that is public - i.e. not a “private address” according to the table here, or “link local” according to the specification here, return that address.
c. Otherwise, return any one of the IPv4 addresses available.